requirements.txt
Flask==2.3.3
Flask-Mail==0.9.1
openai==1.3.8
python-dotenv==1.0.1

